+++ odm2api/ODM2/services/deleteService.py
@@ -1,7 +1,10 @@
 from __future__ import (absolute_import, division, print_function)
+import warnings
+warnings.warn('The module odm2api.ODM2.services.deleteService will be depricated. '
+              'Please use odm2api.services.deleteService instead.',
+              FutureWarning, stacklevel=2)
 
-from odm2api.ODM2 import serviceBase
-from odm2api.ODM2.models import TimeSeriesResultValues
+from odm2api.services.deleteService import DeleteODM2 as newClass
 
 
 __author__ = 'jmeline'
@@ -9,34 +12,8 @@ __author__ = 'jmeline'
 # Annotations
 
 
-class DeleteODM2(serviceBase):
-
-    def remove(self, obj):
-        self._session.delete(obj)
-
-# CV
-# Core
-# Data Quality
-# Equipment
-# Extension Properties
-# External Identifiers
-# Lab Analyses
-# Provenance
-# Annotations
-# Sampling Features
-# Sensors
-# Result Values
-
-    def deleteTSRValues(self, ids=None, startdate=None, dates=None):
-
-        q = self._session.query(TimeSeriesResultValues)
-        if ids:
-            q = q.filter(TimeSeriesResultValues.ResultID.in_(ids))
-        if startdate:
-            # delete all values on or after the startdate.
-            q = q.filter(TimeSeriesResultValues.ValueDateTime >= startdate)
-        if dates:
-            q = q.filter(TimeSeriesResultValues.ValueDateTime.in_(dates))
-        numvals = q.count()
-        q.delete(False)
-        return numvals
+def DeleteODM2(*args, **kwargs):
+    warnings.warn('The class odm2api.ODM2.services.readService.DeleteODM2 will be depricated. '
+                  'Please use odm2api.services.readService.DeleteODM2 instead.',
+                  FutureWarning, stacklevel=2)
+    return newClass(*args, **kwargs)
