+++ odm2api/services/__init__.py
@@ -1,0 +1,15 @@
+from __future__ import (absolute_import, division, print_function)
+
+from odm2api.services.createService import CreateODM2
+from odm2api.services.deleteService import DeleteODM2
+from odm2api.services.readService import ReadODM2
+from odm2api.services.updateService import UpdateODM2
+
+__author__ = 'jmeline'
+
+__all__ = [
+    'CreateODM2',
+    'DeleteODM2',
+    'ReadODM2',
+    'UpdateODM2'
+]
